extract data.js
