+++ lib/data.js
@@ -1,0 +1,24 @@
+/* global document, fetch  */
+
+'use strict';
+
+const Utils = require('./utils');
+
+module.exports.load = function (userId, queryFrom, queryTo) {
+
+    const timeEntriesUrl = `https://app.10000ft.com/api/v1/users/${userId}/time_entries?fields=approvals&from=${Utils.isoDate(queryFrom)}&page=1&per_page=1000&to=${Utils.isoDate(queryTo)}&with_suggestions=true`;
+    const timeEntriesPromise = fetch(timeEntriesUrl, { credentials: 'include' }).then((res) => res.json());
+
+    const projectsUrl = `https://app.10000ft.com/api/v1/users/${userId}/projects?with_archived=true&per_page=100&with_phases=true`;
+    const projectsPromise = fetch(projectsUrl, { credentials: 'include' }).then((res) => res.json());
+
+    const leaveTypesUrl = `https://app.10000ft.com/api/v1/leave_types?page=1&with_archived=true`;
+    const leaveTypesPromise = fetch(leaveTypesUrl, { credentials: 'include' }).then((res) => res.json());
+
+    const existing = document.getElementById('results-3048m');
+    if (existing) {
+        existing.style.opacity = '0.3';
+    }
+
+    return Promise.all([timeEntriesPromise, projectsPromise, leaveTypesPromise]);
+};
