+++ lib/index.js
@@ -1,6 +1,7 @@
-/* global document, fetch, alert */
+/* global document, fetch */
 'use strict';
 
+const Data = require('./data');
 const Templates = require('./templates');
 const Utils = require('./utils');
 
@@ -8,35 +9,6 @@ const now = new Date();
 
 const queryFrom = new Date(now.getFullYear(), now.getMonth() - 1, 1); // from start of previous month
 const queryTo = new Date(now.getFullYear(), now.getMonth() + 2, 0); // till end of next month
-
-const load = () => {
-
-    const userId = window.whoami.id;
-
-    const timeEntriesUrl = `https://app.10000ft.com/api/v1/users/${userId}/time_entries?fields=approvals&from=${Utils.isoDate(queryFrom)}&page=1&per_page=1000&to=${Utils.isoDate(queryTo)}&with_suggestions=true`;
-    const timeEntriesPromise = fetch(timeEntriesUrl, { credentials: 'include' }).then((res) => res.json());
-
-    const projectsUrl = `https://app.10000ft.com/api/v1/users/${userId}/projects?with_archived=true&per_page=100&with_phases=true`;
-    const projectsPromise = fetch(projectsUrl, { credentials: 'include' }).then((res) => res.json());
-
-    const leaveTypesUrl = `https://app.10000ft.com/api/v1/leave_types?page=1&with_archived=true`;
-    const leaveTypesPromise = fetch(leaveTypesUrl, { credentials: 'include' }).then((res) => res.json());
-
-    const existing = document.getElementById('results-3048m');
-    if (existing) {
-        existing.style.opacity = '0.3';
-    }
-
-    Promise.all([timeEntriesPromise, projectsPromise, leaveTypesPromise])
-        .then(([timeEntriesRes, projectsRes, laveTypesRes]) => {
-
-            report({
-                timeEntries: timeEntriesRes.data,
-                projects: projectsRes.data,
-                leaveTypes: laveTypesRes.data
-            });
-        });
-};
 
 const getInitialValue = function () {
 
@@ -260,8 +232,16 @@ const report = ({ timeEntries, projects, leaveTypes }) => {
 };
 
 if (!window.whoami) {
-    alert('Are you on 10kft? And logged in?');
+    window.alert('Are you on 10kft? And logged in?');
     return;
 }
 
-load();
+Data.load(window.whoami.id, queryFrom, queryTo)
+    .then(([timeEntriesRes, projectsRes, laveTypesRes]) => {
+
+        report({
+            timeEntries: timeEntriesRes.data,
+            projects: projectsRes.data,
+            leaveTypes: laveTypesRes.data
+        });
+    });
